nonsmooth diagnostic - supposed to work like this: 1. fail on first mismatched comparison op  in the forward sweep (there is no point in wasting compute cycles to continue just to count how many failures there are since it doesn't tell me if for example they all come from one location in the code or a variety of locations) 2. return the operation code and the operation index (from the count of all operations) in the error message as an exception the user can parse 3. call the newly introduce "IndependentDiag" as alternative trace start that takes the operation index parsed in 2 as the index where the trace should stop at the X point where the forward sweep failed 4. This trace call will now throw an exception when it reaches said operation index and I can scrutinize the call stack to find the exact spot where the offending operation is located in the user code.
This all is to address the problem that a given user code may, for example have many if conditions which never
trigger a bad comparison operator because they are benign but then there may be a few of the many that do.
This mechanism is intended as a tool to find these few among the many if conditions and remedy them appropriately, e.g.
by conditional assignments.